requirements.txt
pydantic>=2.7.0
openai>=1.3.0
groq>=0.4.0
python-dotenv==1.0.0
openenv-core==0.1.0
gradio>=4.0.0